Printing Data Frames

Print a data frame.

# S3 method for data.frame
print(x, ...)

Arguments

x

object of class data.frame.

...

optional arguments to print methods.

Details

This calls format which formats the data frame column-by-column, then converts to a character matrix and dispatches to the print method for matrices.

When quote = TRUE only the entries are quoted not the row names nor the column names.

See also

Examples

(dd <- data.frame(x = 1:8, f = gl(2,4), ch = I(letters[1:8])))
#>   x f ch
#> 1 1 1  a
#> 2 2 1  b
#> 3 3 1  c
#> 4 4 1  d
#> 5 5 2  e
#> 6 6 2  f
#> 7 7 2  g
#> 8 8 2  h
     # print() with defaults
print(dd, quote = TRUE, row.names = FALSE)
#>    x   f  ch
#>  "1" "1" "a"
#>  "2" "1" "b"
#>  "3" "1" "c"
#>  "4" "1" "d"
#>  "5" "2" "e"
#>  "6" "2" "f"
#>  "7" "2" "g"
#>  "8" "2" "h"
     # suppresses row.names and quotes all entries