Markdown Template Properties

These functions can be used in R Markdown documents.

rmarkdown_author(user_id = NULL)

rmarkdown_department()

rmarkdown_date(date = Sys.Date())

rmarkdown_table(...)

rmarkdown_template(type = "latex")

rmarkdown_logo(logo_type = "front")

Arguments

user_id

user ID at Certe, defaults to currently logged in user

date

Date to print in "d mmmm yyyy"

...

data set (and options) to pass on to tbl_flextable()

type

defaults to "latex", must be "latex" or "word" (case-insensitive)

logo_type

type of logo, must be "front" or "footer" (case-insensitive). For the LaTeX template, the front logo must be max 16x7 cm, and the footer logo must be max 16x0.7 cm.

Details

For LaTeX, 4 logo files can be used to add to the resulting PDF file: logocentre (38x58 mm), logoleft (38x58 mm), logoright (38x58 mm) and logofooter (7 mm high). They must be set as YAML parameters in an R Markdown file header.